Fix: Panel markup should not be present.
